# selectively increase the optimization level per module/library