--- a/boards/px4/io-v2/default.cmake
+++ b/boards/px4/io-v2/default.cmake
@@ -4,6 +4,7 @@ px4_add_board(
 	VENDOR px4
 	MODEL io-v2
 	TOOLCHAIN arm-none-eabi
+	CONSTRAINED_FLASH
 	ARCHITECTURE cortex-m3
 	DRIVERS
 	MODULES