--- a/platforms/common/px4_work_queue/CMakeLists.txt
+++ b/platforms/common/px4_work_queue/CMakeLists.txt
@@ -43,4 +43,5 @@ if(PX4_TESTING)
 	add_subdirectory(test)
 endif()
 
+target_compile_options(px4_work_queue PRIVATE ${MAX_CUSTOM_OPT_LEVEL})
 target_link_libraries(px4_work_queue PRIVATE px4_platform)