--- a/platforms/nuttx/src/px4/common/CMakeLists.txt
+++ b/platforms/nuttx/src/px4/common/CMakeLists.txt
@@ -32,7 +32,7 @@
 ############################################################################
 
 # skip for px4_layer support on an IO board
-if (NOT ${PX4_BOARD} MATCHES "px4_io")
+if(NOT PX4_BOARD MATCHES "px4_io")
 
 	add_library(px4_layer
 		board_crashdump.c