--- a/platforms/nuttx/src/px4/stm/stm32_common/dshot/CMakeLists.txt
+++ b/platforms/nuttx/src/px4/stm/stm32_common/dshot/CMakeLists.txt
@@ -34,3 +34,4 @@
 px4_add_library(arch_dshot
 	dshot.c
 )
+target_compile_options(arch_dshot PRIVATE ${MAX_CUSTOM_OPT_LEVEL})