--- a/platforms/nuttx/src/px4/stm/stm32_common/hrt/CMakeLists.txt
+++ b/platforms/nuttx/src/px4/stm/stm32_common/hrt/CMakeLists.txt
@@ -34,5 +34,8 @@
 px4_add_library(arch_hrt
 	hrt.c
 )
-target_compile_options(arch_hrt PRIVATE -Wno-cast-align) # TODO: fix and enable
-
+target_compile_options(arch_hrt
+	PRIVATE
+		${MAX_CUSTOM_OPT_LEVEL}
+		-Wno-cast-align # TODO: fix and enable
+)