--- a/platforms/nuttx/src/px4/stm/stm32_common/io_pins/CMakeLists.txt
+++ b/platforms/nuttx/src/px4/stm/stm32_common/io_pins/CMakeLists.txt
@@ -32,8 +32,9 @@
 ############################################################################
 
 px4_add_library(arch_io_pins
+	input_capture.c
 	io_timer.c
 	pwm_servo.c
 	pwm_trigger.c
-	input_capture.c
 )
+target_compile_options(arch_io_pins PRIVATE ${MAX_CUSTOM_OPT_LEVEL})