--- a/platforms/nuttx/src/px4/stm/stm32_common/spi/CMakeLists.txt
+++ b/platforms/nuttx/src/px4/stm/stm32_common/spi/CMakeLists.txt
@@ -34,3 +34,4 @@
 px4_add_library(arch_spi
 	spi.cpp
 )
+target_compile_options(arch_spi PRIVATE ${MAX_CUSTOM_OPT_LEVEL})