--- a/src/lib/drivers/accelerometer/CMakeLists.txt
+++ b/src/lib/drivers/accelerometer/CMakeLists.txt
@@ -35,9 +35,5 @@ px4_add_library(drivers_accelerometer
 	PX4Accelerometer.cpp
 	PX4Accelerometer.hpp
 )
-target_link_libraries(drivers_accelerometer
-	PRIVATE
-		conversion
-		drivers__device
-		mathlib
-)
+target_compile_options(drivers_accelerometer PRIVATE ${MAX_CUSTOM_OPT_LEVEL})
+target_link_libraries(drivers_accelerometer PRIVATE conversion drivers__device)