--- a/src/lib/drivers/gyroscope/CMakeLists.txt
+++ b/src/lib/drivers/gyroscope/CMakeLists.txt
@@ -35,4 +35,5 @@ px4_add_library(drivers_gyroscope
 	PX4Gyroscope.cpp
 	PX4Gyroscope.hpp
 )
+target_compile_options(drivers_gyroscope PRIVATE ${MAX_CUSTOM_OPT_LEVEL})
 target_link_libraries(drivers_gyroscope PRIVATE conversion drivers__device)