--- a/src/lib/perf/CMakeLists.txt
+++ b/src/lib/perf/CMakeLists.txt
@@ -33,3 +33,4 @@
 
 add_library(perf perf_counter.cpp)
 add_dependencies(perf prebuild_targets)
+target_compile_options(perf PRIVATE ${MAX_CUSTOM_OPT_LEVEL})