--- a/src/modules/mc_rate_control/CMakeLists.txt
+++ b/src/modules/mc_rate_control/CMakeLists.txt
@@ -37,6 +37,7 @@ px4_add_module(
 	MODULE modules__mc_rate_control
 	MAIN mc_rate_control
 	COMPILE_FLAGS
+		${MAX_CUSTOM_OPT_LEVEL}
 	SRCS
 		MulticopterRateControl.cpp
 		MulticopterRateControl.hpp