--- a/src/modules/sensors/vehicle_acceleration/CMakeLists.txt
+++ b/src/modules/sensors/vehicle_acceleration/CMakeLists.txt
@@ -35,6 +35,9 @@ px4_add_library(vehicle_acceleration
 	VehicleAcceleration.cpp
 	VehicleAcceleration.hpp
 )
+
+target_compile_options(vehicle_acceleration PRIVATE ${MAX_CUSTOM_OPT_LEVEL})
+
 target_link_libraries(vehicle_acceleration
 	PRIVATE
 		mathlib