--- a/src/modules/sensors/vehicle_imu/CMakeLists.txt
+++ b/src/modules/sensors/vehicle_imu/CMakeLists.txt
@@ -37,4 +37,5 @@ px4_add_library(vehicle_imu
 	VehicleIMU.cpp
 	VehicleIMU.hpp
 )
+target_compile_options(vehicle_imu PRIVATE ${MAX_CUSTOM_OPT_LEVEL})
 target_link_libraries(vehicle_imu PRIVATE sensor_corrections px4_work_queue)