On an stm32f4 this saves about 4-8% cpu depending on the configuration at a cost of about 16 kB of extra flash.

As a comparison, building the entire system -O2 reduces cpu somewhere around 15% (px4_fmu-v4), but at a cost of several hundred kilobytes of flash (well over the 2MB limit).





### Master (px4_fmu-v4_default SYS_AUTOSTART=4001)

``` Console
 PID COMMAND                   CPU(ms) CPU(%)  USED/STACK PRIO(BASE) STATE FD
   0 Idle Task                   14746 25.616   244/  512   0 (  0)  READY  3
   1 hpwork                          0  0.000   344/ 1260 249 (249)  w:sig  3
   2 lpwork                          0  0.000   344/ 1612  50 ( 50)  w:sig  3
   3 init                         1420  0.000  2064/ 2924 100 (100)  w:sem  3
   4 wq:manager                      1  0.000   408/ 1252 255 (255)  w:sem  4
 440 mavlink_rcv_if0               215  0.362  2520/ 4060 175 (175)  w:sem  4
  15 wq:hp_default                 689  1.233  1072/ 1900 240 (240)  w:sem  4
  19 wq:lp_default                  32  0.000   912/ 1700 205 (205)  w:sem  4
 155 wq:SPI2                       307  0.580   968/ 2492 252 (252)  w:sem  4
 157 wq:SPI1                      9475 17.271  1748/ 2492 253 (253)  w:sem  4
 186 wq:navigation_and_contro    10801 19.883  5072/ 7196 241 (241)  w:sem  4
 188 wq:rate_ctrl                 6729 12.119  1128/ 1660 255 (255)  w:sem  4
 199 commander                     727  1.233  1280/ 3212 140 (140)  w:sig  6
 200 commander_low_prio              1  0.000   692/ 2996  50 ( 50)  w:sem  6
 206 mavlink_if0                  2786  4.862  1832/ 2572 100 (100)  READY  4
 210 mavlink_if1                  2391  4.136  1888/ 2540 100 (100)  READY  4
 212 mavlink_rcv_if1               243  0.435  2536/ 4060 175 (175)  w:sem  4
 265 gps                            47  0.072  1008/ 1676 205 (205)  w:sig  4
 301 mavlink_if2                   913  1.669  1832/ 2484 100 (100)  READY  4
 304 mavlink_rcv_if2               247  0.362  2536/ 4060 175 (175)  w:sem  4
 319 wq:UART4                      387  0.725   752/ 1396 234 (234)  w:sem  4
 561 top                          1463  3.338  1248/ 2028 239 (239)  RUN    3
 453 wq:attitude_ctrl              947  2.031  1060/ 1532 242 (242)  w:sem  4
 469 navigator                      57  0.145  1016/ 1764 105 (105)  w:sem  5
 534 logger                        288  0.507  1872/ 3644 230 (230)  w:sem  3
 535 log_writer_file                 0  0.000   368/ 1164  60 ( 60)  w:sem  3

Processes: 26 total, 5 running, 21 sleeping, max FDs: 15
CPU usage: 70.97% tasks, 3.41% sched, 25.62% idle
DMA Memory: 5120 total, 0 used 0 peak
Uptime: 57.126s total, 14.747s idle

```

### PR (px4_fmu-v4_default SYS_AUTOSTART=4001)

``` Console
 PID COMMAND                   CPU(ms) CPU(%)  USED/STACK PRIO(BASE) STATE FD
   0 Idle Task                   13983 32.173   244/  512   0 (  0)  READY  3
   1 hpwork                          0  0.000   344/ 1260 249 (249)  w:sig  3
   2 lpwork                          0  0.000   344/ 1612  50 ( 50)  w:sig  3
   3 init                         1421  0.000  2064/ 2924 100 (100)  w:sem  3
   4 wq:manager                      1  0.000   400/ 1252 255 (255)  w:sem  4
 440 mavlink_rcv_if0               167  0.362  2480/ 4060 175 (175)  w:sem  4
  15 wq:hp_default                 507  1.231  1224/ 1900 240 (240)  w:sem  4
  19 wq:lp_default                  25  0.072   872/ 1700 205 (205)  w:sem  4
 155 wq:SPI2                       230  0.579   840/ 2492 252 (252)  w:sem  4
 157 wq:SPI1                      6863 15.797  1824/ 2492 253 (253)  w:sem  4
 186 wq:navigation_and_contro     7309 17.173  5080/ 7196 241 (241)  w:sem  4
 188 wq:rate_ctrl                 4468 10.434  1192/ 1660 255 (255)  w:sem  4
 199 commander                     549  1.159  1280/ 3212 140 (140)  READY  6
 200 commander_low_prio              1  0.000   692/ 2996  50 ( 50)  w:sem  6
 206 mavlink_if0                  2156  5.000  1832/ 2572 100 (100)  w:sig  4
 210 mavlink_if1                  1849  4.347  1832/ 2540 100 (100)  w:sig  4
 212 mavlink_rcv_if1               185  0.434  2536/ 4060 175 (175)  w:sem  4
 265 gps                            36  0.072  1016/ 1676 205 (205)  w:sem  4
 301 mavlink_if2                   696  1.594  1832/ 2484 100 (100)  READY  4
 304 mavlink_rcv_if2               188  0.434  2536/ 4060 175 (175)  w:sem  4
 319 wq:UART4                      289  0.652   760/ 1396 234 (234)  READY  4
 555 top                          1335  3.333  1296/ 2028 239 (239)  RUN    3
 453 wq:attitude_ctrl              443  1.231  1280/ 1596 242 (242)  w:sem  4
 469 navigator                      41  0.144  1016/ 1764 105 (105)  READY  5
 534 logger                        213  0.507  1688/ 3644 230 (230)  w:sem  3
 535 log_writer_file                 0  0.000   368/ 1164  60 ( 60)  w:sem  3

Processes: 26 total, 6 running, 20 sleeping, max FDs: 15
CPU usage: 64.57% tasks, 3.26% sched, 32.17% idle
DMA Memory: 5120 total, 0 used 0 peak
Uptime: 44.550s total, 13.984s idle

```